Added info about Danube Cloud release edition into output of GET system_version - closes #21
