+++ core/version.py
@@ -1,1 +1,2 @@
-__version__ = '2.3.2a'
+__version__ = '2.3.2b'
+__edition__ = 'ce'
